Backend-Bot-master/pyproject.toml
[tool.poetry]
name = "poetry-mining"
version = "0.1.0"
description = ""
authors = ["vladgenyuk29 <vlad.genyuk@mail.ru>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
aiohttp = "^3.11.10"
python-dotenv = "^1.0.1"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
fastapi = {extras = ["all"], version = "^0.115.6"}
uvicorn = "^0.32.1"
loguru = "^0.7.3"
sqlalchemy = {version = "^2.0.36", extras = ["asyncio"]}
pyjwt = "^2.10.1"
asyncpg = "^0.30.0"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.25.3"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
matplotlib = "^3.10.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
